requirements.txt
certifi==2015.9.6.2
funcsigs==0.4
influxdb==2.9.3
kafka-python==0.9.4
pbr==1.8.1
profilehooks==1.8.0
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.8.1
six==1.10.0
ujson==1.33
virtualenv==13.1.2
wheel==0.24.0

